Add files via upload
This file contains the sql queries I used to analyze a dataset that contains data about some of the most popular songs on Spotify in 2024. The dataset, "Most Streamed Spotify Songs 2024" by Nidula Elgiriyewithana published on Kaggle, only contains data from January to approximately June of 2024, but it's still useful for practicing analyzing data in sql.